파일 위치 수정 에러수정
