--- src/css/UserInfo.css
@@ -1,17 +1,0 @@
-.user-info{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 50px;
-    padding: 1%;
-    
-}
-
-.clock{
-    font-size: 30px;
-}
-
-.update_button{
-    text-align: center;
-    font-size: 50px;
-}
