+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './css/index.css';
+import './styles/index.css';
 import AppRouter from './AppRouter';
 import reportWebVitals from './reportWebVitals';
