+++ src/styles/App.css
@@ -1,0 +1,104 @@
+@import url('https://fonts.googleapis.com/css2?family=Hi+Melody&family=Nanum+Pen+Script&display=swap');
+
+body {
+  font-family: "Hi Melody", sans-serif;
+  margin: 0; /* 기본 마진 제거 */
+  padding: 0; /* 기본 패딩 제거 */
+  overflow-x: hidden; /* 수평 스크롤 제거 */
+}
+
+.App {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+}
+
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: navy;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.App-clock {
+  background-color: gray;
+  display: flex;
+  color: black;
+  font-size: medium;
+}
+
+.calender {
+  display: inline-block;
+  padding: 1%;
+  text-align: center;
+  height: auto;
+  overflow: hidden;
+}
+
+.calender .react-calendar__navigation__label {
+  font-weight: bold;
+}
+
+.weather {
+  display: inline-block;
+  padding-bottom: 20px;
+  position: relative;
+  height: auto;
+  overflow: hidden;
+}
+
+.calendar-weather-container {
+  display: flex;
+  justify-content: center; /* 중앙 정렬 */
+  align-items: center; /* 요소들을 중앙 정렬 */
+  margin-top: 20px;
+  flex-wrap: wrap; /* 화면 크기에 따라 줄바꿈을 적용 */
+  padding: 10px;
+  width: 100%; /* 전체 너비 설정 */
+}
+
+.calendar-container,
+.weather-widget {
+  margin: 3px; /* 간격 최소화 */
+  flex: 1;
+  min-width: 300px; /* 최소 너비를 설정하여 화면이 좁아졌을 때 줄바꿈 적용 */
+  max-width: 500px; /* 최대 너비 설정 */
+  display: flex;
+  justify-content: center;
+}
+
+.buttons-container {
+  display: flex;
+  justify-content: center; /* 버튼들을 가운데 정렬 */
+  margin-top: 20px;
+  flex-direction: row; /* 버튼들을 수평으로 정렬 */
+  gap: 10px; /* 버튼 간격 조정 */
+}
