+++ src/styles/NavigationBar.css
@@ -1,0 +1,14 @@
+.navbar-todo {
+    background-color: blanchedalmond !important;
+    width: 100%; /* 네비게이션 바를 화면 너비에 맞게 설정 */
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .navbar-item h1 {
+    margin: 0;
+  }
+
