+++ src/styles/index.css
@@ -12,3 +12,7 @@ code {
     monospace;
 }
 
+html, body {
+  margin: 0px;
+  overflow-x: hidden;
+}
